requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
comtypes==1.4.5
exceptiongroup==1.2.2
idna==3.7
iniconfig==2.0.0
numpy==1.24.4
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
psutil==6.0.0
pyperclip==1.9.0
pytest==8.2.2
pywin32==306
pywinauto==0.6.8
PyYAML==6.0.1
Random-Word==1.0.11
requests==2.32.3
six==1.16.0
tomli==2.0.1
urllib3==2.2.2
